feat(ui): implement adaptive layout with burger menu
