+++ src/app/core/i18n/tndm-language-switcher/tndm-language-switcher.scss
@@ -15,7 +15,7 @@
     cursor: pointer;
     color: v.$color-grey;
     font-family: v.$font-family-heading;
-    font-size: 14px;
+    font-size: 12px;
     text-transform: uppercase;
     transition: color 0.2s ease;
