+++ src/app/shared/components/tndm-header/tndm-header.html
@@ -1,0 +1,40 @@
+<header class="header">
+  <tndm-lang-switcher />
+
+  <nav class="navigation" [class.open]="isMenuOpen">
+    <ul class="nav" [class.open]="isMenuOpen">
+      @if(authService.isAuthenticated){ @for (item of navItems; track item.path) {
+      <li class="nav__item">
+        <a class="nav__link" routerLinkActive="nav__link_active" [routerLink]="item.path" (click)="closeMenu()">
+          {{ item.label }}
+        </a>
+      </li>
+      } }
+
+      <li class="nav__auth-mobile">
+        @if(!authService.isAuthenticated){
+        <div class="header__auth-btns">
+          <tndm-button [btnConfig]="signUpBtnConfig" (clicked)="onSignUpClick()" />
+          <tndm-button [btnConfig]="signInBtnConfig" (clicked)="onSignInClick()" />
+        </div>
+        } @else {
+        <tndm-button [btnConfig]="logoutBtnConfig" (clicked)="onLogoutClick()" />
+        }
+      </li>
+    </ul>
+
+    <div class="header__auth-desktop">
+      @if(!authService.isAuthenticated){
+      <div class="header__auth-btns">
+        <tndm-button [btnConfig]="signUpBtnConfig" (clicked)="onSignUpClick()" />
+        <tndm-button [btnConfig]="signInBtnConfig" (clicked)="onSignInClick()" />
+      </div>
+      } @else {
+      <tndm-button [btnConfig]="logoutBtnConfig" (clicked)="onLogoutClick()" /> }
+    </div>
+  </nav>
+
+  <button class="burger-menu" type="button" [class.open]="isMenuOpen" (click)="toggleMenu()">
+    <span></span><span></span><span></span><span></span>
+  </button>
+</header>
