+++ src/app/shared/components/tndm-header/tndm-header.scss
@@ -1,0 +1,162 @@
+@use 'styles/typography' as t;
+@use 'styles/variables' as v;
+@use 'styles/mixins' as m;
+@use 'styles/adaptive' as a;
+
+.navigation {
+  display: flex;
+  align-items: center;
+  width: 100%;
+  justify-content: center;
+
+  @include a.tablet() {
+    justify-content: end;
+  }
+}
+
+.header {
+  position: relative;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  box-sizing: border-box;
+  border: solid v.$color-grey;
+  border-width: 0 1px 1px 1px;
+
+  padding: 10px 15px;
+
+  @include a.tablet() {
+    gap: 20px;
+  }
+
+  @include a.mobile() {
+    padding: 5px 10px;
+    border-width: 0 0 1px;
+
+    &__auth-desktop {
+      display: none;
+    }
+  }
+
+  &__auth-btns {
+    display: flex;
+    gap: 25px;
+    width: fit-content;
+
+    @include a.mobile() {
+      flex-direction: column;
+      gap: 10px;
+    }
+  }
+}
+
+.nav {
+  display: flex;
+  gap: 16px;
+  margin: 0;
+  padding: 0;
+  list-style: none;
+  justify-content: center;
+  @include t.text-body(v.$text-primary-color);
+  font-size: 14px;
+  width: 100%;
+
+  &__auth-mobile {
+    display: none;
+    margin-top: 20px;
+  }
+
+  @include a.tablet() {
+    display: none;
+    width: auto;
+
+    &.open {
+      display: flex;
+      position: absolute;
+      flex-direction: column;
+      right: 20px;
+      top: 5px;
+      padding: 50px 30px 30px;
+      z-index: 20;
+      background-color: v.$bg-surface;
+      border: 1px solid v.$border-second-color;
+    }
+  }
+
+  @include a.mobile() {
+    &__auth-mobile {
+      display: block;
+    }
+  }
+
+  &__item {
+    @include m.corner-borders($color: v.$color-accent);
+    text-align: center;
+    &::after,
+    &::before {
+      opacity: 0;
+    }
+
+    &:hover {
+      &::after,
+      &::before {
+        opacity: 1;
+      }
+      &::after {
+        right: -9px;
+        bottom: -9px;
+      }
+      &::before {
+        top: -9px;
+        left: -9px;
+      }
+    }
+  }
+
+  &__link {
+    color: inherit;
+    text-decoration: none;
+  }
+}
+
+.burger-menu {
+  display: none;
+
+  @include a.tablet() {
+    display: grid;
+    border: none;
+    display: grid;
+    grid-template-columns: repeat(2, 1fr);
+    gap: 6px;
+    padding: 8px;
+    cursor: pointer;
+    position: relative;
+    z-index: 40;
+    background: none;
+    transition:
+      transform 0.2s,
+      gap 0.2s 0.2s;
+
+    span {
+      width: 5px;
+      height: 5px;
+      background: v.$color-accent;
+      justify-self: center;
+      align-self: center;
+      transition: 0.4s;
+    }
+
+    &:hover,
+    &.open {
+      transform: rotate(45deg);
+
+      span {
+        transform: rotate(45deg);
+      }
+    }
+
+    &.open {
+      transform: translateX(-300%);
+    }
+  }
+}
