+++ src/app/shared/components/tndm-main-layout/tndm-main-layout.ts
@@ -2,11 +2,9 @@ import { ChangeDetectionStrategy, Component, inject } from '@angular/core';
 import { ActivatedRoute, NavigationEnd, Router, RouterOutlet } from '@angular/router';
 import { filter, map, startWith } from 'rxjs';
 import { toSignal } from '@angular/core/rxjs-interop';
-import { TndmAuthService } from '@auth';
 import { TndmButton } from '../../ui/tndm-button/tndm-button';
-import { TndmNavigation } from '../tndm-navigation/tndm-navigation';
+import { TndmNavigation } from '../tndm-header/tndm-header';
 import { TndmTitleStrategy } from '../../../core/title-strategy/tndm-title-strategy';
-import { NavigationService } from '../../../core/navigation/navigation.service';
 import { TndmLangSwitcher } from '../../../core/i18n/tndm-language-switcher/tndm-language-switcher';
 
 @Component({
@@ -20,9 +18,6 @@ export class TndmMainLayout {
   private router = inject(Router);
   private route = inject(ActivatedRoute);
   private titleStrategy = inject(TndmTitleStrategy);
-  readonly authService: TndmAuthService = inject(TndmAuthService);
-
-  readonly navService: NavigationService = inject(NavigationService);
 
   readonly pageTitle = this.titleStrategy.pageTitle;
 
@@ -44,26 +39,4 @@ export class TndmMainLayout {
     ),
     { initialValue: { showLines: false, headerType: 'default' } }
   );
-
-  readonly signInBtnConfig = { label: 'sign in' } as const;
-  readonly signUpBtnConfig = {
-    label: 'sign up',
-    variant: 'secondary',
-  } as const;
-  readonly logoutBtnConfig = {
-    label: 'logout',
-    variant: 'secondary',
-  } as const;
-
-  onSignInClick(): void {
-    this.navService.goToLogin();
-  }
-
-  onSignUpClick(): void {
-    this.navService.goToRegister();
-  }
-
-  onLogoutClick(): void {
-    this.navService.logout();
-  }
 }
