--- src/app/shared/components/tndm-navigation/tndm-navigation.scss
@@ -1,45 +1,0 @@
-@use 'styles/typography' as t;
-@use 'styles/variables' as v;
-@use 'styles/mixins' as m;
-
-:host {
-  margin: 0 30px 0 auto;
-}
-
-.nav {
-  display: flex;
-  gap: 20px;
-  width: 100%;
-  margin: 0;
-  padding: 0;
-  list-style: none;
-  @include t.text-body(v.$text-primary-color);
-
-  &__item {
-    @include m.corner-borders($color: v.$color-accent);
-    &::after,
-    &::before {
-      opacity: 0;
-    }
-
-    &:hover {
-      &::after,
-      &::before {
-        opacity: 1;
-      }
-      &::after {
-        right: -9px;
-        bottom: -9px;
-      }
-      &::before {
-        top: -9px;
-        left: -9px;
-      }
-    }
-  }
-
-  &__link {
-    color: inherit;
-    text-decoration: none;
-  }
-}
