+++ src/app/shared/ui/tndm-button/tndm-button.scss
@@ -11,6 +11,7 @@
   border: none;
   padding: 0;
   text-transform: lowercase;
+  white-space: nowrap;
   background-color: inherit;
   transition: transform 0.2s;
   cursor: pointer;
