+++ src/styles/_adaptive.scss
@@ -1,0 +1,23 @@
+@use 'styles/variables' as v;
+
+@mixin mobile {
+  @media (max-width: v.$bp-mobile) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media (max-width: v.$bp-tablet) {
+    @content;
+  }
+}
+
+@mixin tablet {
+  @media (max-width: v.$bp-desktop) {
+    @content;
+  }
+}
+
+@mixin adaptiv-font($pcSize, $mobSize) {
+  font-size: calc(#{$mobSize + px} + #{$pcSize - $mobSize} * ((100vw - 320px) / 960));
+}
